requirements.txt

openai
python-dotenv
streamlit
langchain_openai